login

+++ routes/web.php
@@ -6,6 +6,8 @@ use App\Http\Controllers\backend\ProdukController as BackendProdukController;
 use App\Http\Controllers\backend\ImageController as BackendImageController;
 use App\Http\Controllers\Frontend\HomeController;
 use App\Http\Controllers\Frontend\ProdukController;
+use App\Http\Controllers\LoginController;
+use App\Http\Controllers\AdminController;
 /*
 |--------------------------------------------------------------------------
 | Web Routes
@@ -17,43 +19,62 @@ use App\Http\Controllers\Frontend\ProdukController;
 |
 */
 
-Route::get('/login', function () {
-    return view('dashboard');
+// Route::get('/login', function () {
+//     return view('dashboard');
+// });
+
+
+Route::middleware(['guest'])->group(function () {
+    Route::get('/', [LoginController::class, 'index'])->name('login');
+    Route::post('/login-proses', [LoginController::class, 'login_proses'])->name('login-proses');
 });
 
 
- //BE category
- Route::get('/admin/category', [KategoriController::class, 'index'])->name('beCategory');
- Route::get('admin/category/create', [KategoriController::class, 'create'])->name('beCategori.create');
- Route::post('/admin/category', [KategoriController::class, 'store'])->name('beCategory.store');
-
- Route::get('/admin/kategori-edit/{id}', [KategoriController::class, 'edit'])->name('beCategory.edit');
- Route::put('/admin/kategori-update/{id}', [KategoriController::class, 'update'])->name('beCategory.update');
-
- Route::delete('/admin/kategori-Delete/{id}', [KategoriController::class, 'destroy'])->name('beCategory.destroy');
 
 
- //be Produk
- Route::get('/admin/produk/{id}', [BackendProdukController::class, 'index'])->name('beProduk.index');
- Route::get('/admin/produk-create/{id}', [BackendProdukController::class, 'create'])->name('beProduk.create');
- Route::post('/admin/produk-store', [BackendProdukController::class, 'store'])->name('beProduk.store');
- Route::get('/admin/produk-edit/{id}', [BackendProdukController::class, 'edit'])->name('beProduk.edit');
- Route::put('/admin/produk-update/{id}', [BackendProdukController::class, 'update'])->name('beProduk.update');
- Route::delete('/admin/produk-Delete/{id}', [BackendProdukController::class, 'destroy'])->name('beProduk.destroy');
-
- Route::get('/admin/image/{id}', [BackendImageController::class, 'index'])->name('beImage.index');
- Route::get('/admin/image-create/{id}', [BackendImageController::class, 'create'])->name('beImage.create');
- Route::post('/admin/image-store', [BackendImageController::class, 'store'])->name('beImage.store');
-
- Route::get('/admin/image-edit/{id}', [BackendImageController::class, 'edit'])->name('beImage.edit');
- Route::put('/admin/image-update/{id}', [BackendImageController::class, 'update'])->name('beImage.update');
- Route::delete('/admin/image-Delete/{id}', [BackendImageController::class, 'destroy'])->name('beImage.destroy');
+Route::get('/home', function () {
+    return redirect()->route('admin.dashboard');
+});
 
 
+Route::middleware(['auth'])->group(function () {
+    Route::get('/logout', [LoginController::class, 'logout'])->name('logout');
+
+    //BE category
+    Route::get('/admin', [AdminController::class, 'index'])->name('admin.dashboard')->middleware('userAkses:admin');
+    Route::get('/admin/category', [KategoriController::class, 'index'])->name('beCategory')->middleware('userAkses:admin');
+    Route::get('admin/category/create', [KategoriController::class, 'create'])->name('beCategori.create')->middleware('userAkses:admin');
+    Route::post('/admin/category', [KategoriController::class, 'store'])->name('beCategory.store')->middleware('userAkses:admin');
+
+    Route::get('/admin/kategori-edit/{id}', [KategoriController::class, 'edit'])->name('beCategory.edit')->middleware('userAkses:admin');
+    Route::put('/admin/kategori-update/{id}', [KategoriController::class, 'update'])->name('beCategory.update')->middleware('userAkses:admin');
+
+    Route::delete('/admin/kategori-Delete/{id}', [KategoriController::class, 'destroy'])->name('beCategory.destroy')->middleware('userAkses:admin');
 
 
- Route::resource('/', HomeController::class);
-// FE
-Route::get('/kategori/{id}', [ProdukController::class, 'index'])->name('feProduk');
+    //be Produk
+    Route::get('/admin/produk/{id}', [BackendProdukController::class, 'index'])->name('beProduk.index')->middleware('userAkses:admin');
+    Route::get('/admin/produk-create/{id}', [BackendProdukController::class, 'create'])->name('beProduk.create')->middleware('userAkses:admin');
+    Route::post('/admin/produk-store', [BackendProdukController::class, 'store'])->name('beProduk.store')->middleware('userAkses:admin');
+    Route::get('/admin/produk-edit/{id}', [BackendProdukController::class, 'edit'])->name('beProduk.edit')->middleware('userAkses:admin');
+    Route::put('/admin/produk-update/{id}', [BackendProdukController::class, 'update'])->name('beProduk.update')->middleware('userAkses:admin');
+    Route::delete('/admin/produk-Delete/{id}', [BackendProdukController::class, 'destroy'])->name('beProduk.destroy')->middleware('userAkses:admin');
 
-Route::get('/kategori/show/{id}', [ProdukController::class, 'show'])->name('feShow');
+    Route::get('/admin/image/{id}', [BackendImageController::class, 'index'])->name('beImage.index')->middleware('userAkses:admin');
+    Route::get('/admin/image-create/{id}', [BackendImageController::class, 'create'])->name('beImage.create')->middleware('userAkses:admin');
+    Route::post('/admin/image-store', [BackendImageController::class, 'store'])->name('beImage.store')->middleware('userAkses:admin');
+
+    Route::get('/admin/image-edit/{id}', [BackendImageController::class, 'edit'])->name('beImage.edit')->middleware('userAkses:admin');
+    Route::put('/admin/image-update/{id}', [BackendImageController::class, 'update'])->name('beImage.update')->middleware('userAkses:admin');
+    Route::delete('/admin/image-Delete/{id}', [BackendImageController::class, 'destroy'])->name('beImage.destroy')->middleware('userAkses:admin');
+
+
+    // FE
+    Route::get('/home', [HomeController::class, 'index'])->middleware('userAkses:customer')->name('customer.dashboard');;
+    Route::get('/kategori/{id}', [ProdukController::class, 'index'])->name('feProduk')->middleware('userAkses:customer');
+
+    Route::get('/kategori/show/{id}', [ProdukController::class, 'show'])->name('feShow')->middleware('userAkses:customer');
+
+});
+
+
